requirements.txt
asgiref==3.5.0
mysqlclient==2.1.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
Django==4.0.3
django-cors-headers==3.11.0
django-mysql==4.5.0
djangorestframework==3.13.1
django-redis==5.2.0
Faker==13.3.4
psycopg2==2.9.3
PyJWT==2.3.0